+++ package.json
@@ -14,7 +14,7 @@
     "ts:watch": "ts-node-dev --no-notify --transpileOnly -r esm -r dotenv/config",
     "dev:workspaces": "wsrun --exclude-missing dev",
     "dev:playground": "prisma playground --port 5000 --server-only",
-    "dev": "run-p dev:workspaces dev:playground",
+    "dev": "run-p dev:**",
     "deps": "rimraf 'yarn.lock' 'node_modules' '**/node_modules' && yarn",
     "lint": "eslint '**/*.{ts,tsx}'",
     "lint:fix": "yarn lint --fix",
@@ -38,6 +38,7 @@
     "@typescript-eslint/eslint-plugin": "^1.4.1",
     "@typescript-eslint/parser": "^1.4.1",
     "babel-plugin-module-resolver": "^3.1.2",
+    "babel-plugin-relay": "^3.0.0",
     "dotenv": "^7.0.0",
     "eslint": "^5.14.1",
     "eslint-config-airbnb": "^17.1.0",
@@ -48,6 +49,8 @@
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-react-hooks": "^1.2.0",
     "eslint-plugin-relay": "^1.0.0",
+    "graphql": "^14.1.1",
+    "npm-run-all": "^4.1.5",
     "prisma": "^1.26.6",
     "rimraf": "^2.6.3",
     "ts-node": "^8.0.0",
