--- packages/components/README.md
@@ -1,3 +1,0 @@
-# Notes
-
-TODO: Replace react-native-web with react-native when we add it.
