+++ packages/relay/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@app/relay",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "relay": "relay-compiler --src '../' --schema '../api/schema.graphql' --language typescript --artifactDirectory 'generated'",
+    "dev": "yarn relay --watch",
+    "gen": "yarn relay --watchman false"
+  },
+  "dependencies": {
+    "relay-runtime": "^3.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@types/react-relay": "^1.3.11",
+    "@types/relay-runtime": "^1.3.12",
+    "graphql": "^14.1.1",
+    "relay-compiler-language-typescript": "^2.0.0",
+    "relay-compiler": "^3.0.0"
+  }
+}
