+++ packages/web/package.json
@@ -3,11 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "relay": "relay-compiler --src . --include 'components/**' 'pages/**' --schema '../api/schema.graphql' --language typescript --artifactDirectory 'generated'",
-    "gen": "yarn relay --watchman false",
-    "dev:next": "next",
-    "dev:relay": "yarn relay --watch",
-    "dev": "run-p dev:**",
+    "dev": "next",
     "build": "next build",
     "start": "next start",
     "analyze": "BUNDLE_ANALYZE=both next build",
@@ -18,6 +14,7 @@
     "@app/api": "1.0.0",
     "@app/components": "1.0.0",
     "@app/hooks": "1.0.0",
+    "@app/relay": "1.0.0",
     "@app/themes": "1.0.0",
     "@sentry/browser": "^4.5.3",
     "cookie": "^0.3.1",
@@ -26,36 +23,31 @@
     "next": "^8.0.0",
     "nprogress": "^0.2.0",
     "prop-types": "^15.7.2",
-    "react": "^16.8.1",
     "react-art": "^16.8.0",
     "react-dom": "^16.8.1",
     "react-intl": "^2.8.0",
     "react-native-web": "^0.11.2",
     "react-relay": "^3.0.0",
+    "react": "^16.8.1",
     "relay-runtime": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0-0",
+    "@babel/core": "^7.0.0",
     "@types/cookie": "^0.3.2",
     "@types/gravatar": "^1.4.28",
     "@types/next": "^8.0.0",
     "@types/nprogress": "0.0.29",
-    "@types/react": "^16.7.11",
     "@types/react-dom": "^16.0.11",
     "@types/react-intl": "^2.3.14",
     "@types/react-native": "^0.57.41",
     "@types/react-relay": "^1.3.11",
+    "@types/react": "^16.7.11",
     "@types/relay-runtime": "^1.3.12",
     "@types/styled-jsx": "^2.2.7",
     "@zeit/next-bundle-analyzer": "^0.1.2",
     "@zeit/next-typescript": "^1.1.1",
-    "babel-plugin-relay": "^3.0.0",
-    "graphql": "^14.1.1",
     "next-plugin-custom-babel-config": "^1.0.0",
     "next-transpile-modules": "^2.0.0",
-    "npm-run-all": "^4.1.5",
-    "relay-compiler": "^3.0.0",
-    "relay-compiler-language-typescript": "^2.0.0",
     "utility-types": "^3.4.1"
   }
 }
