+++ packages/web/pages/index.tsx
@@ -4,10 +4,10 @@ import { Text, View } from 'react-native';
 import { createFragmentContainer, graphql } from 'react-relay';
 import useAppContext from '@app/hooks/useAppContext';
 import usePageTitles from '@app/hooks/usePageTitles';
-import CreateWeb from '../components/CreateWeb';
-import Layout from '../components/Layout';
-import Link from '../components/Link';
-import { pagesQuery } from '../generated/pagesQuery.graphql';
+import CreateWeb from '@app/components/CreateWeb';
+import Layout from '@app/components/Layout';
+import Link from '@app/components/Link';
+import { pagesQuery } from '@app/relay/generated/pagesQuery.graphql';
 
 type Viewer = NonNullable<pagesQuery['viewer']>;
