+++ tsconfig.json
@@ -16,5 +16,5 @@
     "strict": true,
     "target": "esnext"
   },
-  "exclude": ["node_modules", "**/generated/**/*.*"]
+  "exclude": ["node_modules", "**/api/generated/**/*.*"]
 }
